fix: fix model product long to decimal

+++ Controllers/ProductController.cs
@@ -34,7 +34,7 @@ public class ProductController : ControllerBase
         public string name { get; set; }
         public string description { get; set; }
 
-        public long price { get; set; }
+        public decimal price { get; set; }
 
         public int stock { get; set; }
